requirements.txt
pyserial==3.5
pymodbus==3.5.2
